+++ azure-pipelines.yml
@@ -40,10 +40,10 @@ jobs:
       matrix:
         node-12:
           node_version: ^12.0.0
-          webpack_version: 5.0.0-beta.0
+          webpack_version: 5.0.0-beta.2
         node-10:
           node_version: ^10.13.0
-          webpack_version: 5.0.0-beta.0
+          webpack_version: 5.0.0-beta.2
         node-10-v4:
           node_version: ^10.13.0
           webpack_version: 4
@@ -88,10 +88,10 @@ jobs:
       matrix:
         node-12:
           node_version: ^12.0.0
-          webpack_version: 5.0.0-beta.0
+          webpack_version: 5.0.0-beta.2
         node-10:
           node_version: ^10.13.0
-          webpack_version: 5.0.0-beta.0
+          webpack_version: 5.0.0-beta.2
         node-10-v4:
           node_version: ^10.13.0
           webpack_version: 4
@@ -136,10 +136,10 @@ jobs:
       matrix:
         node-12:
           node_version: ^12.0.0
-          webpack_version: 5.0.0-beta.0
+          webpack_version: 5.0.0-beta.2
         node-10:
           node_version: ^10.13.0
-          webpack_version: 5.0.0-beta.0
+          webpack_version: 5.0.0-beta.2
         node-10-v4:
           node_version: ^10.13.0
           webpack_version: 4
