+++ src/utils/ready.js
@@ -4,6 +4,9 @@ export default function ready(context, fn, req) {
     return fn(context.stats);
   }
 
-  context.log.info(`wait until bundle finished: ${req.url || fn.name}`);
+  const name = req.url || fn.name;
+
+  context.log.info(`wait until bundle finished${name ? `: ${name}` : ''}`);
+
   context.callbacks.push(fn);
 }
