+++ test/fixtures/simple-config/webpack.config.js
@@ -1,11 +1,12 @@
 'use strict';
 
+const path = require('path');
+
 module.exports = {
   mode: 'development',
   context: __dirname,
   entry: './foo.js',
   output: {
-    path: '/',
-  },
-  node: false,
+    path: path.resolve(__dirname, '../../outputs/'),
+  }
 };
